% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest_weekly_expirations.R
\name{backtest_weekly_expirations}
\alias{backtest_weekly_expirations}
\title{Backtest weekly expirations}
\usage{
backtest_weekly_expirations(start_date, end_date)
}
\arguments{
\item{start_date}{date}

\item{end_date}{date}
}
\value{
dataframe
}
\description{
Returns a dataframe that consists of all weekly expriations
in date range, along with execution dates.  This is used to organize
all the expirations in a backtest.
}
