% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{following_friday}
\alias{following_friday}
\title{Following Friday}
\usage{
following_friday(dt)
}
\arguments{
\item{dt}{date}
}
\value{
date
}
\description{
If input date is not a Friday, returns the following Friday
}
