% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expiration.R
\name{is_third_friday}
\alias{is_third_friday}
\title{Check if third Friday of month}
\usage{
is_third_friday(dt_friday)
}
\arguments{
\item{dt_friday}{date}
}
\value{
boolean
}
\description{
This function checks if the given Friday is the
third Friday of the month
}
