% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expiration.R
\name{monthly_expiration}
\alias{monthly_expiration}
\title{Regular monthly expiration}
\usage{
monthly_expiration(year, month)
}
\arguments{
\item{year}{integer}

\item{month}{integer}
}
\value{
date
}
\description{
Use this function to get the regular expiration for a given year month.
}
