% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expiration.R
\name{monthly_last_tradeday}
\alias{monthly_last_tradeday}
\title{Last trade-date for monthly expiration}
\usage{
monthly_last_tradeday(year, month)
}
\arguments{
\item{year}{integer}

\item{month}{integer}
}
\description{
Use this function to get the last trade date for the regular-expiration
option for this year month.
}
