% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade_day.R
\name{nearest_tradeday}
\alias{nearest_tradeday}
\title{Nearest business day}
\usage{
nearest_tradeday(dt)
}
\arguments{
\item{current_date}{date}
}
\value{
date
}
\description{
If the current date is a business day, returns the current date.
Otherwise return the next business day.
}
