% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expiration.R
\name{next_monthly_expiration}
\alias{next_monthly_expiration}
\title{Find the next monthly expiration}
\usage{
next_monthly_expiration(dt)
}
\arguments{
\item{dt}{date}
}
\value{
date
}
\description{
For the given input date, will find the next expiration that's after the
end-of-day of the input date
}
