% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{previous_friday}
\alias{previous_friday}
\title{Previous Friday}
\usage{
previous_friday(dt)
}
\arguments{
\item{dt}{date}
}
\value{
date
}
\description{
If input date is not a Friday, returns the previous Friday
}
