% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expiration.R
\name{third_friday}
\alias{third_friday}
\title{Third Friday of month}
\usage{
third_friday(year, month)
}
\arguments{
\item{year}{integer}

\item{month}{integer}
}
\value{
date
}
\description{
Use this function to get the date of the third Friday
of the given year and month.  This is an internal function.
}
