% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly_expirations_2002_2020.R
\docType{data}
\name{weekly_expirations_2002_2020}
\alias{weekly_expirations_2002_2020}
\title{Weekly expirations from the beginning of 2002 to the end of 2020
Used for organizing backtests of weekly option trading strategies}
\format{
A dataframe
}
\usage{
weekly_expirations_2002_2020
}
\description{
Weekly expirations from the beginning of 2002 to the end of 2020
Used for organizing backtests of weekly option trading strategies
}
\keyword{datasets}
