% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeday.R
\name{nearest_tradeday}
\alias{nearest_tradeday}
\title{Nearest trade date}
\usage{
nearest_tradeday(dt)
}
\arguments{
\item{dt}{date}
}
\value{
date
}
\description{
If the current date is a trade date, returns the current date.
Otherwise return the next business day.
}
